+++ smartech_app/lib/app_inbox/widgets/smt_simple_notification_view.dart
@@ -1,7 +1,8 @@
-import 'dart:collection';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:smartech_app/app_inbox/utils/utils.dart';
+import 'package:smartech_app/navigator.dart';
+import 'package:smartech_app/update_profile.dart';
 import 'package:smartech_appinbox/model/smt_appinbox_model.dart';
 import 'package:url_launcher/url_launcher.dart';
 
@@ -51,10 +52,12 @@ class SMTSimpleNotificationView extends StatelessWidget {
                                           final Uri _url = Uri.parse(e.actionDeeplink);
                                           if (!await launchUrl(_url)) throw 'Could not launch $_url';
                                           // await FlutterWebBrowser.openWebPage(url: e.actionDeeplink);
+                                        } else if (e.actionDeeplink.contains("smartechflutter://profile")) {
+                                          NavigationUtilities.pushRoute(UpdateProfile.route);
                                         } else {
-                                          Map<String, dynamic> dict = HashMap();
-                                          dict["actionDeeplink"] = e.actionDeeplink;
-                                          dict['isFromScreen'] = true;
+                                          // Map<String, dynamic> dict = HashMap();
+                                          // dict["actionDeeplink"] = e.actionDeeplink;
+                                          // dict['isFromScreen'] = true;
                                           // NavigationUtilities.pushRoute(
                                           //   DeepLinkScreen.route,
                                           //   args: dict,
@@ -68,7 +71,8 @@ class SMTSimpleNotificationView extends StatelessWidget {
                                     )
                                   : e.aTyp == 2
                                       ? InkWell(
-                                          onTap: () {
+                                          onTap: () async {
+                                            print("navigation called");
                                             Clipboard.setData(ClipboardData(text: e.configCtxt)).then((result) {
                                               final snackBar = SnackBar(
                                                 content: Text('Copied'),
@@ -76,6 +80,14 @@ class SMTSimpleNotificationView extends StatelessWidget {
                                               );
                                               ScaffoldMessenger.of(context).showSnackBar(snackBar); // -> show a notification
                                             });
+                                            if (e.actionDeeplink.contains("http")) {
+                                              print("navigate to browser with url");
+                                              final Uri _url = Uri.parse(e.actionDeeplink);
+                                              if (!await launchUrl(_url)) throw 'Could not launch $_url';
+                                              // await FlutterWebBrowser.openWebPage(url: e.actionDeeplink);
+                                            } else if (e.actionDeeplink.contains("smartechflutter://profile")) {
+                                              NavigationUtilities.pushRoute(UpdateProfile.route);
+                                            }
                                           },
                                           child: Text(
                                             e.actionName.toString(),