+++ smartech_app/lib/navigator.dart
@@ -66,7 +66,7 @@ class NavigationUtilities {
 /// arguments for the screen.
 Route<dynamic>? onGenerateRoute(RouteSettings settings) {
   final routeName = settings.name;
-  final arguments = settings.arguments as Map<String, dynamic>;
+  final arguments = settings.arguments as Map<dynamic, dynamic>;
   // final routeType = arguments["routeType"] as RouteType? ?? RouteType.defaultRoute;
 
   navigator(Widget screen) {